refactor: core structure
